move test_sort macro to meta crate
