+++ meta/Cargo.toml
@@ -5,6 +5,7 @@ version = "0.1.0"
 [dependencies]
 hyper = "0.9"
 lazy_static = "0.2.1"
+rand = "0.3"
 regex = "0.1"
 toml = "0.1"
 url = "1.1.1"
